Refactored the download link functionality into a separate component
