+++ src/views/Exercises/Show/Downloads.vue
@@ -18,14 +18,10 @@
           Job Description Template
         </dt>
         <dd class="govuk-summary-list__value">
-          <a
-            id="job-description-template"
-            class="govuk-link"
-            :class="{'download-visited': documentsDownloaded.jobDescription}"
-            href="javascript:void(0)"
-            @click.prevent="download(exercise.uploadedJobDescriptionTemplate); documentsDownloaded.jobDescription = true; "
-          >
-            {{ exercise.uploadedJobDescriptionTemplate }}</a>
+          <DownloadLink
+            :file-name="exercise.uploadedJobDescriptionTemplate"
+            :exercise-id="exerciseId"
+          />
         </dd>
       </div>
       <div class="govuk-summary-list__row">
@@ -33,13 +29,10 @@
           Terms and Conditions Template
         </dt>
         <dd class="govuk-summary-list__value">
-          <a
-            id="terms-and-conditions-template"
-            class="govuk-link"
-            :class="{'download-visited': documentsDownloaded.termsAndConditions}"
-            href="javascript:void(0)"
-            @click.prevent="download(exercise.uploadedTermsAndConditionsTemplate); documentsDownloaded.jobDescription = true;"
-          >{{ exercise.uploadedTermsAndConditionsTemplate }}</a>
+          <DownloadLink
+            :file-name="exercise.uploadedTermsAndConditionsTemplate"
+            :exercise-id="exerciseId"
+          />
         </dd>
       </div>
       <div class="govuk-summary-list__row">
@@ -47,13 +40,10 @@
           Independent Assessor Template
         </dt>
         <dd class="govuk-summary-list__value">
-          <a
-            id="independent-assessor-template"
-            class="govuk-link"
-            :class="{'download-visited': documentsDownloaded.independentAssessor}"
-            href="javascript:void(0)"
-            @click="download(exercise.uploadedIndependentAssessorTemplate); documentsDownloaded.independentAssessor = true;"
-          >{{ exercise.uploadedIndependentAssessorTemplate }}</a>
+          <DownloadLink
+            :file-name="exercise.uploadedIndependentAssessorTemplate"
+            :exercise-id="exerciseId"
+          />
         </dd>
       </div>
       <div class="govuk-summary-list__row">
@@ -61,13 +51,10 @@
           Candidate Assessment Form Template
         </dt>
         <dd class="govuk-summary-list__value">
-          <a
-            id="candidate-assessment-form-template"
-            class="govuk-link"
-            :class="{'download-visited': documentsDownloaded.candidateAssessmentForm}"
-            href="javascript:void(0)"
-            @click="download(exercise.uploadedCandidateAssessmentFormTemplate); documentsDownloaded.independentAssessor = true; "
-          >{{ exercise.uploadedCandidateAssessmentFormTemplate }}</a>
+          <DownloadLink
+            :file-name="exercise.uploadedCandidateAssessmentFormTemplate"
+            :exercise-id="exerciseId"
+          />
         </dd>
       </div>
     </dl>
@@ -75,18 +62,11 @@
 </template>
 
 <script>
-import firebase from 'firebase';
+import DownloadLink from '@/components/DownloadLink';
 
 export default {
-  data () {
-    return {
-      documentsDownloaded: {
-        jobDescription: false,
-        termsAndConditions: false,
-        independentAssessor: false,
-        candidateAssessmentForm: false,
-      },
-    };
+  components: {
+    DownloadLink,
   },
   computed: {
     exercise() {
@@ -99,59 +79,9 @@ export default {
       return this.$store.getters['exerciseDocument/id'];
     },
   },
-  methods: {
-    download(fileName) {
-      // Create a reference to the file we want to download
-      const fileSavePath = `exercise-${this.exerciseId}/${fileName}`;
-
-      // Get a reference to the storage service, which is used to create references in your storage bucket
-      const storage = firebase.storage();
-
-      // Create a storage reference from our storage service
-      const storageRef = storage.ref();
-
-      // Create a reference with an initial file path and name
-      const fileNameRef = storageRef.child(fileSavePath);
-
-      // Get the download URL
-      fileNameRef.getDownloadURL().then((url) => {
-        // open url in another window
-        window.open(url);
-      }).catch((error) => {
-
-        // A full list of error codes is available at
-        // https://firebase.google.com/docs/storage/web/handle-errors
-        switch (error.code) {
-        case 'storage/object-not-found':
-          // File doesn't exist
-          break;
-
-        case 'storage/unauthorized':
-          // User doesn't have permission to access the object
-          break;
-
-        case 'storage/canceled':
-          // User canceled the upload
-          break;
-
-        case 'storage/unknown':
-          // Unknown error occurred, inspect the server response
-          break;
-        }
-      });
-    },
-    lookVisited(id) {
-      var download = document.getElementById(id);
-      download.classList.toggle('download-visted');
-    },
-  },
 };
 </script>
 
 <style scoped>
-
-.download-visited {
-  color: #4c2c92;
-}
 
 </style>
